updated table structure

+++ src/main/java/com/jpdr/apps/demo/webflux/authentication/repository/LoginUserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-public interface LoginUserRepository extends ReactiveCrudRepository<LoginUser, Integer> {
+public interface LoginUserRepository extends ReactiveCrudRepository<LoginUser, Long> {
   
   Flux<LoginUser> findAllByIsActiveIsTrue();
   Mono<LoginUser> findByUsernameAndIsActiveIsTrue(String username);
